Expose connection class object to NetworkTemplate
